Bump entities to u128 to avoid collisions (#117) (#393)

+++ crates/bevy_ecs/hecs/src/archetype.rs
@@ -38,7 +38,7 @@ pub struct Archetype {
     types: Vec<TypeInfo>,
     state: HashMap<TypeId, TypeState>,
     len: u32,
-    entities: Box<[u32]>,
+    entities: Box<[u128]>,
     // UnsafeCell allows unique references into `data` to be constructed while shared references
     // containing the `Archetype` exist
     data: UnsafeCell<NonNull<u8>>,
@@ -226,16 +226,16 @@ impl Archetype {
     }
 
     #[allow(missing_docs)]
-    pub fn iter_entities(&self) -> impl Iterator<Item = &u32> {
+    pub fn iter_entities(&self) -> impl Iterator<Item = &u128> {
         self.entities.iter().take(self.len as usize)
     }
 
     #[inline]
-    pub(crate) fn entities(&self) -> NonNull<u32> {
+    pub(crate) fn entities(&self) -> NonNull<u128> {
         unsafe { NonNull::new_unchecked(self.entities.as_ptr() as *mut _) }
     }
 
-    pub(crate) fn entity_id(&self, index: u32) -> u32 {
+    pub(crate) fn entity_id(&self, index: u32) -> u128 {
         self.entities[index as usize]
     }
 
@@ -263,7 +263,7 @@ impl Archetype {
 
     /// # Safety
     /// Every type must be written immediately after this call
-    pub unsafe fn allocate(&mut self, id: u32) -> u32 {
+    pub unsafe fn allocate(&mut self, id: u128) -> u32 {
         if self.len as usize == self.entities.len() {
             self.grow(self.len.max(self.grow_size));
         }
@@ -341,7 +341,7 @@ impl Archetype {
     }
 
     /// Returns the ID of the entity moved into `index`, if any
-    pub(crate) unsafe fn remove(&mut self, index: u32) -> Option<u32> {
+    pub(crate) unsafe fn remove(&mut self, index: u32) -> Option<u128> {
         let last = self.len - 1;
         for ty in &self.types {
             let removed = self
@@ -380,7 +380,7 @@ impl Archetype {
         &mut self,
         index: u32,
         mut f: impl FnMut(*mut u8, TypeId, usize, bool, bool),
-    ) -> Option<u32> {
+    ) -> Option<u128> {
         let last = self.len - 1;
         for ty in &self.types {
             let moved = self
