+++ crates/bevy_scene/src/scene_spawner.rs
@@ -8,7 +8,7 @@ use thiserror::Error;
 use uuid::Uuid;
 
 struct InstanceInfo {
-    entity_map: HashMap<u32, bevy_ecs::Entity>,
+    entity_map: HashMap<u128, bevy_ecs::Entity>,
 }
 
 #[derive(Copy, Clone, Debug, Eq, PartialEq, Hash)]
